backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
ffmpeg-python==0.2.0
sentence-transformers==2.6.1
chromadb==0.4.24
langchain==0.1.0
langchain-community==0.0.13
google-generativeai==0.3.2
python-dotenv==1.0.0
aiofiles==23.2.1 